Create Josephus.java
Introduction
This problem goes back to Josephus Flavius, a 1st century Roman historian
41 rebels are trapped in a cave, surrounded by enemy troops.They decide to commit suicide: they line up in a circle and systematically kill every other one, going around and around, until only one rebel is left -- who supposedly kills himself  (fat chance). 

Task
Given the total number of persons n and a number k which indicates that k-1 persons are skipped and kth person is killed in circle. 
The task is to choose the place in the initial circle so that you are the last one remaining and so survive.

Examples

n = 5 and k = 2 

The safe position is 3. Firstly, the person at position 2 is killed, then person at position 4 is killed, then person at position 1 is killed. Finally, the person at position 5 is killed. So the person at position 3 survives

n = 7 and k = 3

The safe position is 4. The persons at positions 3, 6, 2, 7, 5, 1 are killed in order, and person at position 4 survives